fpn implementation and move to necks
